improved algorithm for mapping/unmapping
In previous versions, all the previous current visible menus are
unmapped and the new currently visible menus are mapped when the
current visible menu changed. Now, using an algorithm that finds
the lowest common ancestor between current and current_new menus,
only the necessary menus are remapped on the display.