Making geom local to main().
I moved all the geometry related code to calcmenu().

The pro is that geom is now a local variable rather than global;
and that all geometry-related code is now in one place.

The con is that it weakens performance a little bit since some
variables, such as the border, have to be reset after the window
creation.